requirements.txt
flask==3.1.3
flask-cors==6.0.2
python-dotenv==1.0.1
requests==2.31.0
instagrapi==2.1.2
